写部分Lab的ui

+++ EdGameFramework/src/main/java/com/edplan/framework/MContext.java
@@ -112,7 +112,7 @@ public class MContext
 		}
 	}
 	
-	public void runOnUIThread(Runnable r,int delayMs){
+	public void runOnUIThread(Runnable r,double delayMs){
 		getUiLooper().post(r,delayMs);
 	}
